server/requirements.txt
Django>=3.1,<3.2
django-cors-headers
boto3
django-storages
ariadne>=0.12,<0.13
psycopg2-binary
pytest-cov
pytest-django
Pillow
PyJWT
whitenoise
django-polymorphic-tree
loguru